adding templates for posts
