+++ blog/urls.py
@@ -3,5 +3,6 @@ from django.urls import path
 from blog import views
 
 urlpatterns = [
-	path('', views.PostView.as_view(), name='home')
+	path("", views.PostView.as_view(), name="home"),
+	path("<slug:slug>/", views.PostDetail.as_view(), name="post_detail"),
 ]
