+++ blog/views/__init__.py
@@ -1,1 +1,2 @@
 from .post_view import PostView
+from .post_view import PostDetail
