--- mysite/blog/urls.py
@@ -1,6 +1,0 @@
-from django.urls import path
-from mysite.blog import views
-
-urlpatterns = [
-    path('', views.PostView.as_view(), name='home')
-]
