--- mysite/blog/views/__init__.py
@@ -1,1 +1,0 @@
-from .post_view import PostView
