--- mysite/blog/views/post_view.py
@@ -1,8 +1,0 @@
-from django.http import HttpResponse
-from django.views import generic
-
-
-class PostView(generic.View):
-
-    def get(self, request, *args, **kwargs):
-        return HttpResponse('Hello World')
